requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
boto3==1.17.21
botocore==1.20.22
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
decorator==5.0.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.3.1
django-classy-tags==2.0.0
django-etc==1.3.0
django-filter==2.4.0
django-froala-editor==3.2.2
django-hitcount==1.3.3
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-storages==1.11.1
django-taggit==1.4.0
django-taggit-serializer==0.1.7
django-taggit-templatetags2==1.6.1
django-templated-mail==1.1.1
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
dynaconf==3.1.4
fake-useragent==0.1.11
feedparser==6.0.2
firebase-admin==5.0.2
google-api-core==1.31.2
google-api-python-client==2.20.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.1
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
googlefinance==0.7
grpcio==1.40.0
httplib2==0.19.1
hvac==0.10.9
idna==2.10
importlib-metadata==2.1.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
knox==0.1.11
loguru==0.5.3
lxml==4.6.2
MarkupSafe==1.1.1
msgpack==1.0.2
multitasking==0.0.9
mysqlclient==2.0.3
numpy==1.20.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==21.0
pandas==1.2.3
pandas-datareader==0.9.0
parse==1.19.0
Pillow==8.1.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==8.1.0
PyJWT==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.4
sgmllib3k==1.0.0
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2
sqlparse==0.4.1
tqdm==4.59.0
uritemplate==3.0.1
urllib3==1.26.3

validators==0.18.2
w3lib==1.22.0
websockets==8.1
yahoo-fin==0.8.7
yfinance==0.1.55
zipp==3.4.0
